requirements.txt
Django==2.2.24
django-jet==1.0.8
djangorestframework==3.11.2
psycopg2-binary==2.8.4
Pillow==8.4.0